requirements.txt
argon2-cffi==21.1.0
asgiref==3.5.2
boto3==1.18.63
botocore==1.21.65
braintree==4.13.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==3.1.7
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-js-asset==2.0.0
django-storages==1.12.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
Markdown==3.3.4
MarkupSafe==2.1.1
mutagen==1.45.1
mysqlclient==1.4.2.post1
oauthlib==3.2.0
Pillow==8.1.2
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stripe==2.58.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==5.3.0
zipp==3.8.1
